grep functionality implemented
To accomplish this, Test-Driven-Development was utilized. A test was written for the search functionality that I know would fail, but is the desired functionality. The search method was written to satisfy this test. The search method iterates over the contents passed to it by the main run method. If the line contains the query passed in to the search method, the line is pushed to an empty vector. In the actual run command