Case Sensitive/Insensitive Search Added
Enviroment variable support was added to the program. An 'ignore case' property was added to the Config struct. Env was imported from standard lib. In the config constructor, if there are no errors, an environment variable is created for IGNORE_CASE. Now, in the run method, depending on the value of the environment variale, the appropriate method is called repsectively.